# Optimize startup times by using `ubrk_clone` instead of `ubrk_open`.