This decreases editor startup time by ~0.35s (8%).

It should be possible to safely cherry-pick this back to 4.4.

## Explanation 

I have previously profiled that `ubrk_open` was causing substantial delays in editor startup times (#102129).
I've dug a bit deeper into the code, and found that cloning `UBreakIterator` instances instead of creating them from scratch doesn't parse the locale (rules?) again (which was the slow part). So, in the interest of speed, I've implemented a method to cache a blueprint `UBreakIterator` for each locale, and only clone it when we need a new one. 

This is not only faster than #102129, but also better in terms of RAM use — because not every `FontAdvanced` instance needs their own `UBreakIterator` caches. Instead, we only have one for each locale globally.

## Benchmarks
I benchmarked a ~0.35s (8%) decrease in editor startup time:

```
 ❯ hyperfine -m25 -iw1 "bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/empty-game --editor --quit" "bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/empty-game --editor --quit"
Benchmark 1: bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/empty-game --editor --quit
  Time (mean ± σ):      4.781 s ±  0.120 s    [User: 3.266 s, System: 0.406 s]
  Range (min … max):    4.469 s …  4.974 s    25 runs

Benchmark 2: bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/empty-game --editor --quit
  Time (mean ± σ):      4.435 s ±  0.095 s    [User: 3.011 s, System: 0.387 s]
  Range (min … max):    4.250 s …  4.557 s    25 runs

Summary
  bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/empty-game --editor --quit ran
    1.08 ± 0.04 times faster than bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/empty-game --editor --quit
```