@AThousandShips 2025-03-23
I don't know about cherry picking a pure enhancement, it isn't a massive performance boost but always exposes possible errors