@Ivorforce 2025-03-23
That's true.
Personally I'd cherry pick it knowing the risks, but I'd be totally ok not to cherry pick it. I just wanted to make the suggestion by adding the label, I hope that's not going against the grain of how these labels are usually applied?