@clayjohn 2025-03-27
> That's true. Personally I'd cherry pick it knowing the risks, but I'd be totally ok not to cherry pick it. I just wanted to make the suggestion by adding the label, I hope that's not going against the grain of how these labels are usually applied?

That's totally fine, PR authors and/or maintainers should definitely apply the cherry pick label as a suggestion for what they think should be cherry picked. In this case, I agree with ATS that this introduces risk without a large enough benefit to justify the risk in a stable release (4.4.2 will be expected to be extremely stable). 

That said, I will leave the cherry pick label, as I expect this is the sort of thing where we will see very quickly whether it causes a regression or not. So if we go to do the 4.4.2 merge batch in a few weeks and this hasn't caused a regression, it is probably safe to include