@AThousandShips 2025-03-27
Absolutely seconding that we should apply the label as we feel is appropriate, thank you for clarifying that and I should have been more clear it was me expressing concerns over doing it and not the adding of the label