@Ivorforce 2025-03-27
It's all good,I just wasn't sure in the moment since I'm still pretty new to this.
Thank you both for clearing that up!